+++ app/spike/internal/cmd/cmd.go
@@ -9,6 +9,10 @@ import (
 	"os"
 
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
+
+	"github.com/spiffe/spike/app/spike/internal/cmd/policy"
+	"github.com/spiffe/spike/app/spike/internal/cmd/secret"
+	"github.com/spiffe/spike/app/spike/internal/cmd/system"
 )
 
 // Initialize sets up the CLI command structure with a workload API X.509
@@ -18,7 +22,7 @@ import (
 //   - get: Retrieves secrets with optional version specification
 //   - delete: Removes specified versions of secrets
 //   - undelete: Restores specified versions of secrets
-//   - init: Initializes the secret management system
+//   - initialization: Initializes the secret management system
 //   - put: Stores new secrets
 //   - list: Displays available secrets
 //
@@ -35,31 +39,9 @@ import (
 //	source := workloadapi.NewX509Source(...)
 //	Initialize(source)
 func Initialize(source *workloadapi.X509Source) {
-	getCmd := NewGetCommand(source)
-	getCmd.Flags().IntP("version", "v", 0, "Specific version to retrieve")
-	rootCmd.AddCommand(getCmd)
-
-	deleteCmd := NewDeleteCommand(source)
-	deleteCmd.Flags().StringP("versions", "v", "0",
-		"Comma-separated list of versions to delete")
-	rootCmd.AddCommand(deleteCmd)
-
-	undeleteCmd := NewUndeleteCommand(source)
-	undeleteCmd.Flags().StringP("versions", "v", "0",
-		"Comma-separated list of versions to undelete")
-	rootCmd.AddCommand(undeleteCmd)
-
-	initCmd := NewInitCommand(source)
-	rootCmd.AddCommand(initCmd)
-
-	//loginCmd := NewLoginCommand(source)
-	//rootCmd.AddCommand(loginCmd)
-
-	putCmd := NewPutCommand(source)
-	rootCmd.AddCommand(putCmd)
-
-	listCmd := NewListCommand(source)
-	rootCmd.AddCommand(listCmd)
+	rootCmd.AddCommand(policy.NewPolicyCommand(source))
+	rootCmd.AddCommand(secret.NewSecretCommand(source))
+	rootCmd.AddCommand(system.NewSystemInitCommand(source))
 }
 
 // Execute runs the root command and handles any errors that occur.