+++ hack/spire-agent-starter.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-#    \\ SPIKE: Secure your secrets with SPIFFE.
-#  \\\\\ Copyright 2024-present SPIKE contributors.
-# \\\\\\\ SPDX-License-Identifier: Apache-2.0
-
-./hack/spire-agent-start.sh
-